refactor: move widget components to seperate folder
