+++ src/widgets/LinkWidget/LinkWidget.tsx
