+++ src/widgets/PlaceholderWidget.tsx
@@ -1,0 +1,3 @@
+export default function PlaceholderWidget() {
+  return <div className="bg-yellow-500">Hello</div>;
+}
