+++ src/widgets/SearchWidget.tsx
