+++ src/widgets/TimeWidget.tsx
